Disable attach button when owner or repo has changes
