--- src/pages/projectSettings/tabs/GeneralTab/Fields/utils.ts
@@ -1,25 +1,0 @@
-const formContext = {};
-
-export const defaultFieldProps = {
-  schema: {},
-  uiSchema: {},
-  idSchema: null,
-  formData: null,
-  errorSchema: {},
-  onChange: () => {},
-  onBlur: () => {},
-  onFocus: () => {},
-  registry: {
-    fields: {},
-    widgets: {},
-    definitions: {},
-    formContext,
-    rootSchema: {},
-  },
-  formContext,
-  autofocus: false,
-  disabled: false,
-  readonly: false,
-  required: false,
-  name: "testField",
-};
